* upstream/master: (93 commits)
  Use /sys/proc/... to get the cdrom device
  [#45112063] - Exec returns an result object not a string, fix expectations accordingly
  Remove placeholder spec accidentally committed
  Use better key for global cache
  Conditionally save compiled packages to global blobstore
  First pass at saving to global blobstore
  [finishes #45168899] AWS bootstrap deploys microbosh from latest CI stemcell
  blobstore client should be able to suggest an object id
  copy global one-ring-key to local path for use by bosh
  CI Publish AMI sends light stemcell to S3
  Catch HTTPClient::ReceiveTimeoutError when waiting for registry to start
  output rds to a consistent file name
  [fixes #45112063] - fix issue with agent startup
  fix stemcell version comparsion when using timestamped stemcells from CI
  Filter for EBS volumes that aren't already deleted or deleting [fixes #45174599]
  support suggesting object id to blobstore server
  [#45168899] Jenkins publishes stemcell AMIs to a well-defined S3 URLs
  [#45168899] AWS bootstrap does micro deploy
  AWS::bootstrap: setting up for a 'no-arg' bosh aws bootstrap micro  command
  Clean up bat dns tests
  ...

+++ blobstore_client/lib/blobstore_client/s3_blobstore_client.rb
@@ -61,10 +61,10 @@ module Bosh
       end
 
       # @param [File] file file to store in S3
-      def create_file(file)
+      def create_file(object_id, file)
         raise BlobstoreError, "unsupported action" if @simple
 
-        object_id = generate_object_id
+        object_id ||= generate_object_id
 
         file = encrypt_file(file) if @encryption_key
 
@@ -136,8 +136,10 @@ module Bosh
       # @param [String] oid object id
       # @return [void]
       def store_in_s3(path, oid)
+        s3_object = get_object_from_s3(oid)
+        raise BlobstoreError, "object id #{oid} is already in use" if s3_object.exist?
         File.open(path, "r") do |temp_file|
-          @s3.buckets[bucket_name].objects[oid].write(temp_file)
+          s3_object.write(temp_file)
         end
       end
